+++ nequip/model/__init__.py
@@ -1,4 +1,4 @@
-from ._eng import EnergyModel
+from ._eng import EnergyModel, SimpleIrrepsConfig
 from ._grads import ForceOutput
 from ._scaling import RescaleEnergyEtc, PerSpeciesRescale
 from ._weight_init import uniform_initialize_FCs
@@ -6,6 +6,7 @@ from ._weight_init import uniform_initialize_FCs
 from ._build import model_from_config
 
 __all__ = [
+    "SimpleIrrepsConfig",
     "EnergyModel",
     "ForceOutput",
     "RescaleEnergyEtc",